+++ Products/CMFCore/utils.py
@@ -238,8 +238,8 @@ def _mergedLocalRoles(object):
             object = object.aq_parent
             object = getattr(object, 'aq_inner', object)
             continue
-        if hasattr(object, 'im_self'):
-            object = object.im_self
+        if hasattr(object, '__self__'):
+            object = object.__self__
             object = getattr(object, 'aq_inner', object)
             continue
         break